# api/swagger.yml
swagger: "2.0"
info:
  description: The Jobber service handles web requests for job reporting functionality
  title: Jobber
  version: 1.0.0
consumes:
- application/json
- text/plain
produces:
- application/json
- text/plain
schemes:
- http
  #- https
paths:
...
  /v1/jobs/{id}/messages:
    parameters:
          - $ref: "#/parameters/IDParam"
    get:
      tags: ["job"]
      operationId: getJobMessages
      responses:
        '200':
          description: "list of job messagess"
          schema:
            type: array
            items:
              $ref: "#/definitions/JobMessage"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    post:
      tags: ["job"]
      operationId: addJobMessage
      consumes:
      - application/json
      parameters:
      - name: "message"
        description: "Creates a new job message"
        required: true
        in: body
        schema:
          type: string
      responses:
        '202':
          description: Accepted
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
definitions:
...
  JobMessage:
    type: object
    required:
    - message
    - created_at
    properties:
      message:
        type: string
      created_at:
        type: string
        format: date-time
...
  Error:
    type: object
    required:
    - code
    - message
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string
parameters:
  IDParam:
    name: id
    type: string
    in: path
    required: true